Moved event system to another package
Other system modules will interact with event system. Therefore
it is better to have it as an intermediate module.
